==> 03a_data_wrangling_main_bar.ipynb <==
These codes attempt to re-simulate the main-bar reviews that the reviewer would have seen whilst writing his/her own review. Please refer to the manuscript to understand how the hypothesized reversed-engineer algorithm works. 

## Importing packages and reading data

from tqdm import tqdm
import pickle
from datetime import datetime, timedelta
import time
import math
import ast
import numpy as np
import pandas as pd
import types
from joblib import Parallel, delayed

def imports():
    for name, val in globals().items():
        if isinstance(val, types.ModuleType):
            try:
                yield val.__name__, val.__version__
                pass
            except:
                yield val.__name__
                pass
            pass
        pass
    pass

list(imports())

df=pd.read_csv("../data/raw_data/FPS_reviews.csv.zip", low_memory=False)

# Convert columns to their supposed data types (for safe measure)
df['recommendation_id'] = df['recommendation_id'].astype(str)
df['game_id'] = df['game_id'].astype(str)
df['review'] = df['review'].astype(str)
df['score'] = df['score'].astype(float)
df['timestamp_created'] = df['timestamp_created'].astype(int)
df['timestamp_updated'] = df['timestamp_updated'].astype(int)
df['votes_up'] = df['votes_up'].astype(int)
df['voted_up'] = df['voted_up'].astype(int)

# organizing data per game

print(f"Removing games with <30 reviews (10 for mainbar, 10 for sidebar, 10 for control).")
game_length=len(set(list(df["game_id"])))
print(f"Before removing games with fewer than 30 reviews, there were {len(df)} reviews from {game_length} games")
df = df.groupby('game_id').filter(lambda x: len(x) >= 30)
df['index'] = [i for i in range(len(df))]
df = df.set_index('index')
game_length=len(set(list(df["game_id"])))
print(f"After  removing games with fewer than 30 reviews, there were {len(df)} reviews from {game_length} games")

df = (df.drop_duplicates(subset=['recommendation_id', 'game_id'], keep='first')).reset_index(drop=True)
game_length=len(set(list(df["game_id"])))
print(f"After removing duplicates, there were {len(df)} reviews from {game_length} games")

df.to_csv("../data/raw_data/data_filtered.csv.gz",compression="gzip",index=False)

## Implementing our hypothesized main-bar algorithm and performing our re-simulation. 



def sort_and_rearrange(temp_df):
    temp_df = temp_df.copy()
    temp_df["score"] = (temp_df["score"]).astype(float)
    temp_df = (temp_df.sort_values("score", ascending=False)).reset_index(drop=True)
    return temp_df

def get_x_days_ago(end_int, days):
    end = datetime.fromtimestamp(end_int)
    start = end - timedelta(days=days)
    start_int = int(time.mktime(start.timetuple()))
    return start_int

def get_x_days_ago_reviews(days, df, end_int):
    start = get_x_days_ago(int(end_int), days)
    df=df.copy()
    df['timestamp_created'] = (df['timestamp_created']).astype(int)
    temp_dates = df[df['timestamp_created'].between(start, end_int - 1)]
    return temp_dates

def condition_against(temp1, temp2, difference):
    if len(temp1) == 10:
        temp = temp1
    elif difference == 10:
        temp = temp2[:len(temp2)]
    elif len(temp2) == 0:
        temp = temp1
    else:
        temp = temp1
        for i in range(difference):
            if i >= len(list(temp2["score"])):
                continue
            if float(list(temp2["score"])[i]) > float(list(temp1["score"])[0]):
                to_be_added = temp2[i:i + 1]
                temp = pd.concat([temp, to_be_added])
    return temp

def temp_conditional_return_1(time_created, day_range1, day_range_ultimate, df):
    day_range1_reviews = get_x_days_ago_reviews(day_range1, df, time_created)
    day_range1_date = get_x_days_ago(time_created, day_range1)
    days_diff = day_range_ultimate - day_range1
    in_between_days_reviews = get_x_days_ago_reviews(days_diff, df, day_range1_date)
    difference = 10 - len(day_range1_reviews)
    temp1 = sort_and_rearrange(day_range1_reviews)
    temp2 = sort_and_rearrange(in_between_days_reviews)
    temp = condition_against(temp1, temp2, difference)
    return temp

def temp_conditional_return_2(time_created, day_range1, day_range2, day_range_ultimate, df):
    day_range1_reviews = get_x_days_ago_reviews(day_range1, df, time_created)
    day_range1_date = get_x_days_ago(time_created, day_range1)
    days_diff_1 = day_range2 - day_range1
    day_range2_reviews = get_x_days_ago_reviews(days_diff_1, df, day_range1_date)
    day_range2_date = get_x_days_ago(time_created, day_range2)
    difference = 10 - len(day_range1_reviews)
    temp1 = sort_and_rearrange(day_range1_reviews)
    temp2 = sort_and_rearrange(day_range2_reviews)
    temp = condition_against(temp1, temp2, difference)
    difference_2 = 10 - len(temp)
    days_diff_2 = day_range_ultimate - day_range2
    day_range3_reviews = get_x_days_ago_reviews(days_diff_2, df, day_range2_date)
    temp3 = sort_and_rearrange(day_range3_reviews)
    temp = condition_against(temp, temp3, difference_2)
    return temp

def temp_conditional_return_3(time_created, day_range1, day_range2, day_range_3, start_date, df):
    day_range1_reviews = get_x_days_ago_reviews(day_range1, df, time_created)
    day_range1_date = get_x_days_ago(time_created, day_range1)
    days_diff_1 = day_range2 - day_range1
    day_range2_reviews = get_x_days_ago_reviews(days_diff_1, df, day_range1_date)
    day_range2_date = get_x_days_ago(time_created, day_range2)
    difference = 10 - len(day_range1_reviews)
    temp1 = sort_and_rearrange(day_range1_reviews)
    temp2 = sort_and_rearrange(day_range2_reviews)
    temp = condition_against(temp1, temp2, difference)
    difference_2 = 10 - len(temp)
    days_diff_2 = day_range_3 - day_range2
    day_range3_reviews = get_x_days_ago_reviews(days_diff_2, df, day_range2_date)
    temp3 = sort_and_rearrange(day_range3_reviews)
    temp = condition_against(temp, temp3, difference_2)
    day_range3_date = get_x_days_ago(time_created, day_range_3)
    day_range_4 = int((time_created - start_date) / 86400)
    difference_3 = 10 - len(temp)
    days_diff_3 = day_range_4 - day_range_3
    day_range4_reviews = get_x_days_ago_reviews(days_diff_3, df, day_range3_date)
    temp4 = sort_and_rearrange(day_range4_reviews)
    temp = condition_against(temp, temp4, difference_3)
    return temp

def get_list_of_main_bar_reviews_old(df):
    order_of_visibility, order_of_scores, order_of_times = [], [], []
    for i in tqdm(range(len(df))):
        game_id = str(df["game_id"][i])
        time_created = int(df["timestamp_created"][i])
        temp = df[(df['game_id'] == game_id) & (df['votes_up'] != 0)]
        temp_dates = get_x_days_ago_reviews(30, temp, time_created)
        if len(temp) == 0 or len(temp_dates) >= 10:
            temp = temp_dates
        else:
            temp_90_dates = temp_conditional_return_1(time_created, 30, 90, temp)
            if len(temp_90_dates) >= 10:
                temp = temp_90_dates
            else:
                temp_180_dates = temp_conditional_return_2(time_created, 30, 90, 180, temp)
                if len(temp_180_dates) >= 10:
                    temp = temp_180_dates
                else:
                    start_date = int(temp['timestamp_created'].min())
                    temp = temp_conditional_return_3(time_created, 30, 90, 180, start_date, temp)

        temp["score"] = pd.to_numeric(temp["score"])
        temp = (temp.sort_values("score", ascending=False)).reset_index(drop=True)

        n = min(len(temp), 10)
        list_of_main_bar_reviews = list(temp["recommendation_id"][:n])
        list_of_scores = list(temp["score"][:n])
        list_of_times = list(temp["timestamp_created"][:n])
        order_of_visibility.append(list_of_main_bar_reviews)
        order_of_scores.append(list_of_scores)
        order_of_times.append(list_of_times)

    return order_of_visibility, order_of_scores, order_of_times

def get_list_of_main_bar_reviews(df):
    ordered_array = np.zeros(shape=(len(df), 10, 3), dtype=float)
    ordered_array[:] = np.nan
    i = np.where(np.isnan(ordered_array[:, 0, 0]))[0][0]
    for i in tqdm(range(i, len(df)), ncols=100):
        game_id = str(df["game_id"][i])
        time_created = int(df["timestamp_created"][i])
        temp = df[(df['game_id'] == game_id) & (df['votes_up'] != 0)]
        temp_dates = get_x_days_ago_reviews(30, temp, time_created)
        if len(temp) == 0 or len(temp_dates) >= 10:
            temp = temp_dates
            pass
        else:
            temp_90_dates = temp_conditional_return_1(time_created, 30, 90, temp)
            if len(temp_90_dates) >= 10:
                temp = temp_90_dates
                pass
            else:
                temp_180_dates = temp_conditional_return_2(time_created, 30, 90, 180, temp)
                if len(temp_180_dates) >= 10:
                    temp = temp_180_dates
                    pass
                else:
                    start_date = int(temp['timestamp_created'].min())
                    temp = temp_conditional_return_3(time_created, 30, 90, 180, start_date, temp)
                    pass
                pass
            pass
        temp["score"] = pd.to_numeric(temp["score"])
        temp = (temp.sort_values("score", ascending=False))
        n = min(len(temp), 10)
        ordered_array[i, :n] = temp[["recommendation_id", "score", "timestamp_created"]][:n]
        pass
    return ordered_array

def get_list_of_main_bar_reviews(df):
    ordered_array = np.zeros(shape=(len(df), 10, 5), dtype=float)
    ordered_array[:] = np.nan
    for i in tqdm(range(len(df.index))):
        game_id = str(df.loc[df.index[i], "game_id"])
        time_created = int(df.loc[df.index[i], "timestamp_created"])
        temp = df[(df['game_id'] == game_id) & (df['votes_up'] != 0)]
        temp_dates = get_x_days_ago_reviews(30, temp, time_created)
        if len(temp) == 0 or len(temp_dates) >= 10:
            temp = temp_dates
            pass
        else:
            temp_90_dates = temp_conditional_return_1(time_created, 30, 90, temp)
            if len(temp_90_dates) >= 10:
                temp = temp_90_dates
                pass
            else:
                temp_180_dates = temp_conditional_return_2(time_created, 30, 90, 180, temp)
                if len(temp_180_dates) >= 10:
                    temp = temp_180_dates
                    pass
                else:
                    start_date = int(temp['timestamp_created'].min())
                    temp = temp_conditional_return_3(time_created, 30, 90, 180, start_date, temp)
                    pass
                pass
            pass
        temp["score"] = pd.to_numeric(temp["score"])
        temp = (temp.sort_values("score", ascending=False))
        n = min(len(temp), 10)
        ordered_array[i, :n, :3] = temp[["recommendation_id", "score", "timestamp_created"]][:n]
        ordered_array[i, :, 3] = np.arange(10)+1
        ordered_array[i, :, 4] = df.index[i]
        pass
    return ordered_array

grouped_df = df.groupby(['game_id'])
split_df = [grouped_df.get_group(x) for x in grouped_df.groups]
ordered_array_split = Parallel(n_jobs=20)(delayed(get_list_of_main_bar_reviews)(split_df[xxx]) for xxx in tqdm(range(len(split_df))))
ordered_array = np.vstack(ordered_array_split)
np.save('../data/temp_reconstruction.npy', ordered_array)

temp_tester_array = ordered_array[~np.any(np.isnan(ordered_array[:, -1, :3]), axis=1)]
unique_reviews_main_bar = [len(np.unique(temp_tester_array[xxx, :, 0])) for xxx in range(temp_tester_array.shape[0])]
unique_reviews_main_bar = np.array(unique_reviews_main_bar)
pd.Series(unique_reviews_main_bar).value_counts(normalize=True).round(3)

### additional testing

## DOESNT WORK
verification_df = df[df['game_id'] == '1000410'].copy()
verification_df["game_id"].index # index messed up, function dependent on index
#verification_df["order_of_main_bar_reviews"], verification_df["order_of_scores"], verification_df["order_of_times"] = get_list_of_main_bar_reviews_old(verification_df)

## Works now
verification_df = df[df['game_id'] == '1000410'].copy()
verification_df=verification_df.reset_index(drop=True)
verification_df["order_of_main_bar_reviews"], verification_df["order_of_scores"], verification_df["order_of_times"] = get_list_of_main_bar_reviews_old(verification_df)

for xyz in tqdm(verification_df.index, ncols=100):
    new_array = ordered_array[ordered_array[:, :, -1] == xyz][:, 0]
    new_array = new_array[~np.isnan(new_array)].astype(int).astype(str)
    old_array = np.array(verification_df.loc[xyz, 'order_of_main_bar_reviews'])
    if not (np.all(old_array == new_array)):
        print('Problem with {}'.format(xyz))
        pass
    pass

for xyz in tqdm(verification_df.index, ncols=100):
    new_array = ordered_array[ordered_array[:, :, -1] == xyz][:, 1]
    new_array = new_array[~np.isnan(new_array)]
    old_array = np.array(verification_df.loc[xyz, 'order_of_scores'])
    if len(old_array) > 0:
        if max(abs(new_array - old_array)) > .001:
            print('Problem with {}'.format(xyz))
            pass
        pass
    pass

for xyz in tqdm(verification_df.index, ncols=100):
    new_array = ordered_array[ordered_array[:, :, -1] == xyz][:, 2]
    new_array = new_array[~np.isnan(new_array)].astype(int)
    old_array = np.array(verification_df.loc[xyz, 'order_of_times'])
    if not (np.all(old_array == new_array)):
        print('Problem with {}'.format(xyz))
        pass
    pass

verification_df = df[df['game_id'] == '559620'].copy().reset_index(drop=True)
verification_df["order_of_main_bar_reviews"], verification_df["order_of_scores"], verification_df["order_of_times"] = get_list_of_main_bar_reviews_old(verification_df)
verification_df.index = df[df['game_id'] == '559620'].index

for xyz in tqdm(verification_df.index, ncols=100):
    new_array = ordered_array[ordered_array[:, :, -1] == xyz][:, 0]
    new_array = new_array[~np.isnan(new_array)].astype(int).astype(str)
    old_array = np.array(verification_df.loc[xyz, 'order_of_main_bar_reviews'])
    if not (np.all(old_array == new_array)):
        print('Problem with {}'.format(xyz))
        pass
    pass

for xyz in tqdm(verification_df.index, ncols=100):
    new_array = ordered_array[ordered_array[:, :, -1] == xyz][:, 1]
    new_array = new_array[~np.isnan(new_array)]
    old_array = np.array(verification_df.loc[xyz, 'order_of_scores'])
    if len(old_array) > 0:
        if max(abs(new_array - old_array)) > .001:
            print('Problem with {}'.format(xyz))
            pass
        pass
    pass

for xyz in tqdm(verification_df.index, ncols=100):
    new_array = ordered_array[ordered_array[:, :, -1] == xyz][:, 2]
    new_array = new_array[~np.isnan(new_array)].astype(int)
    old_array = np.array(verification_df.loc[xyz, 'order_of_times'])
    if not (np.all(old_array == new_array)):
        print('Problem with {}'.format(xyz))
        pass
    pass

#df["order_of_main_bar_reviews"], df["order_of_scores"], df["order_of_times"] = [np.nan]*3

#df.loc[df['game_id'] == x, ["order_of_main_bar_reviews", "order_of_scores", "order_of_times"]]

#df["order_of_main_bar_reviews"], df["order_of_scores"], df["order_of_times"]

## Organizing our data

reviews = pd.DataFrame({'recommendation_id': df["recommendation_id"]})
reviews["review"] = df["review"]

# Removing re-simulations with less than 10 main-bar reviews for consistency in analysis which will be used later.
game_length=len(set(list(df["game_id"])))
print(f"Before removing reviews with <10 default sorted reviews, we have a total of {len(df)} reviews across {game_length} games")
index_of_less_than_10 = []
for i in tqdm(range(len(df))):
    if len((df["order_of_main_bar_reviews"][i])) < 10:
        index_of_less_than_10.append(i)

# df1 = df1.drop(index_of_less_than_10)
df=(df.drop(index_of_less_than_10)).reset_index(drop=True)
game_length=len(set(list(df["game_id"])))
print(f"After removing reviews with <10 default sorted reviews, we have a total of {len(df)} reviews across {game_length} games")

## Saving our data

df.to_csv(f'../data/interim_data/03_data_wrangling/reviews_with_main_bar_reviews.csv.zip', index=False)
