==> 03b_data_wrangling_sidebar_reviews.ipynb <==
import pandas as pd
import ast
import pickle
import types
import concurrent.futures
import numpy as np
from collections import defaultdict
import concurrent.futures
from joblib import Parallel, delayed
from tqdm.auto import tqdm

def imports():
    for name, val in globals().items():
        if isinstance(val, types.ModuleType):
            try:
                yield val.__name__, val.__version__
                pass
            except:
                yield val.__name__
                pass
            pass
        pass
    pass

list(imports())

# Read the files
data_location = '../data/'
main_bar_reviews=np.load(f"{data_location}temp_reconstruction.npy")

reviews_list=pd.read_csv("../data/raw_data/data_filtered.csv.gz", compression="gzip",low_memory=False)

# Convert columns to their supposed data types (for safe measure)
reviews_list['recommendation_id'] = reviews_list['recommendation_id'].astype(str)
reviews_list['game_id'] = reviews_list['game_id'].astype(str)
reviews_list['review'] = reviews_list['review'].astype(str)
reviews_list['score'] = reviews_list['score'].astype(float)
reviews_list['timestamp_created'] = reviews_list['timestamp_created'].astype(int)
reviews_list['timestamp_updated'] = reviews_list['timestamp_updated'].astype(int)
reviews_list['votes_up'] = reviews_list['votes_up'].astype(int)
reviews_list['voted_up'] = reviews_list['voted_up'].astype(int)
reviews_list

# placing main_bar reviews in the dataframe
unique_dict = {}

for subarray in tqdm(main_bar_reviews):
    for row in subarray:
        key = int(row[-1])
        if key not in unique_dict:
            unique_dict[key] = []
        try:
            unique_dict[key].append(str(int(row[0])))
        except:
            unique_dict[key].append(np.nan)
            
reviews_list["order_of_main_bar_reviews"]=(list(unique_dict.values()))

# removing games with <10 main bar reviews
df = (reviews_list[~reviews_list['order_of_main_bar_reviews'].apply(lambda x: np.nan in x)]).reset_index(drop=True)
game_length=len(set(list(df["game_id"])))
print(f"After removing games in which there were <10 main bar reviews, we get {len(df)} across {game_length} games.")

def get_temp_df(df, reviews_list):
    '''
    Generates a dataframe with of column dedicated to sidebar reviews.
    Args:
    df: DataFrame containing the main data.
    reviews_list: DataFrame containing the reviews data.
    
    Returns:
    dataframe with of column dedicated to sidebar reviews.
    '''
    df=df.reset_index(drop=True)
    reviews_list=reviews_list.reset_index(drop=True)
    # Compute start_timestamp for all rows in df
    df['start_timestamp'] = df['timestamp_created'] - (30 * 86400)
    
    # Ensure game_id is of the same type in both DataFrames to avoid type mismatch issues
    df['game_id'] = df['game_id'].astype(str)
    reviews_list['game_id'] = reviews_list['game_id'].astype(str)
    
    # Initialize the ordered array
    ordered_array = np.full((len(df), 10), np.nan, dtype=float)
    unique_game_ids = df['game_id'].unique()
    filtered_reviews_by_game = {
        game_id: reviews_list[(reviews_list['game_id'] == game_id) &
                              (reviews_list['timestamp_created'] > df[df['game_id'] == game_id]['start_timestamp'].min()) &
                              (reviews_list['timestamp_created'] <= df[df['game_id'] == game_id]['timestamp_created'].max())]
        for game_id in tqdm(unique_game_ids)
    }

    for i, row in tqdm(df.iterrows()):
        game_id = row['game_id']
        if game_id in filtered_reviews_by_game:
            relevant_reviews = filtered_reviews_by_game[game_id]
            start_timestamp, time_created = row['start_timestamp'], row['timestamp_created']
            exclude_ids = set(row['order_of_main_bar_reviews'] + [row['recommendation_id']])
            top_reviews = relevant_reviews[(relevant_reviews['timestamp_created'] > start_timestamp) &
                                           (relevant_reviews['timestamp_created'] <= time_created) &
                                           (~relevant_reviews['recommendation_id'].isin(exclude_ids))]
            top_reviews = top_reviews.sort_values('timestamp_created', ascending=False).head(10)['recommendation_id'].values
            
            ordered_array[i, :len(top_reviews)] = top_reviews
    df["sidebar_order_reviews"]=list(ordered_array)
    return df


# Converting data fields to lists
if type(df["order_of_main_bar_reviews"][0])!=list:
    df["order_of_main_bar_reviews"] = [ast.literal_eval(item) for item in df["order_of_main_bar_reviews"]]
    df["order_of_times"] = [ast.literal_eval(item) for item in df["order_of_times"]]
else:
    pass


# running and testing old code here. 

# f1=get_temp_df(df[:1000], reviews_list)
# f1=f1["sidebar_order_reviews"]
# def get_temp_df(df, reviews_list):
#     """
#     Generates a list of unobserved reviews for the past 30 days.

#     Args:
#     df: DataFrame containing the main data.
#     reviews_list: DataFrame containing the reviews data.

#     Returns:
#     A list of unobserved reviews.
#     """
#     side_bar_reviews_list = []
#     for i in tqdm(range(len(df))):
#         current_id=list(df["recommendation_id"])[i]
#         end_int = int(df.timestamp_created[i])
#         start = end_int - (30 * 86400)
#         game_id = str(df.game_id[i])

#         temp_dates = reviews_list[reviews_list['game_id'] == game_id]
#         temp_dates = temp_dates[temp_dates['timestamp_created'].between(start, end_int - 1)]
#         temp_dates = temp_dates.sort_values("timestamp_created", ascending=False).reset_index(drop=True)
#         main_bar_list = df.order_of_main_bar_reviews[i]
#         temp_dates_list = list(temp_dates.recommendation_id)
#         exclude_ids = main_bar_list + [current_id]
#         sidebar_temp_dates_list = temp_dates[~temp_dates['recommendation_id'].isin(exclude_ids)]['recommendation_id'].tolist()
#         if len(temp_dates) >= 10:
            
#             side_bar_reviews_list.append(sidebar_temp_dates_list[:10])
#         else:
#             side_bar_reviews_list.append([])
        
#         # sidebar_temp_dates_list = [x for x in temp_dates_list if x not in main_bar_list+[current_id]]
#         # side_bar_reviews_list.append(sidebar_temp_dates_list[:10])
#     return side_bar_reviews_list

# test_2= get_temp_df(df[:1000], reviews_list)

# # Function to process the array
# def process_array(arr):
#     result = []
#     for subarr in tqdm(arr):
# #         print(subarr)
# #         break
#         if np.isnan(subarr).any():
#             result.append([])
#         else:
#             # Convert each number to string, removing ".0" for whole numbers
#             formatted_subarr = [f"{x:.0f}" if x.is_integer() else str(x) for x in subarr]
#             result.append(formatted_subarr)
#     return result

# # Process the given array
# processed_list = process_array(f1)

# test_2==processed_list


# Setup for parallel processing
grouped_df = df.groupby('game_id')
group_review_list = reviews_list.groupby('game_id')
common_game_ids = set(grouped_df.groups.keys()) & set(group_review_list.groups.keys())
df_chunks = Parallel(n_jobs=20)(
    delayed(get_temp_df)(grouped_df.get_group(game_id), group_review_list.get_group(game_id))
    for game_id in tqdm(common_game_ids)
)


df = pd.concat(df_chunks, ignore_index=True).reset_index(drop=True)
df = (df[~df["sidebar_order_reviews"].apply(lambda x: any(pd.isna(y) for y in x))]).reset_index(drop=True)
game_length=len(set(list(df["game_id"])))
print(f"After removing games in which there were <10 side-bar reviews, we get {len(df)} across {game_length} games.")

df["sidebar_order_reviews"] = (df["sidebar_order_reviews"].apply(
    lambda x: [str(int(y)) if y.is_integer() else str(y) for y in x if not pd.isna(y)]
))

df

csv_file_path = f'{data_location}/interim_data/03_data_wrangling/reviews_with_main_and_sidebar.csv.gz'  
df.to_csv(csv_file_path, index=False, compression='gzip')
