==> 05_preparing_data_for_analysis.ipynb <==
import pandas as pd
from tqdm import tqdm
import types 
import ast
import zipfile
def imports():
    for name, val in globals().items():
        if isinstance(val, types.ModuleType):
            try:
                yield val.__name__, val.__version__
                pass
            except:
                yield val.__name__
                pass
            pass
        pass
    pass

list(imports())

data_location="../data/interim_data/04_text_mining/postprocessed/"
df=pd.read_csv(f'{data_location}data_with_similarities.csv.gz', compression='gzip',low_memory=False)

# adding additional rows to trace which game is a package, which is a package and has identical views compare to the pure game, and which has a package but the package and game view is different. 
df['appears_both_in_game_and_package'] = df.duplicated(subset=["recommendation_id"], keep=False)
df['package_and_game_view_is_identical'] = df.duplicated(subset=["recommendation_id","sidebar_list_of_similarities","main_bar_list_of_similarities","control_list_of_similarities"], keep=False)
df['is_a_package'] = df['game_id'].apply(lambda x: ',' in x)
df["package_review_but_displayed_different"] = (~df["package_and_game_view_is_identical"]) & (df["is_a_package"])

df["appears_both_in_game_and_package"].value_counts()

df["package_and_game_view_is_identical"].value_counts()

df["package_review_but_displayed_different"].value_counts()

df["is_a_package"].value_counts()

# number of package reviews that appeared exactly the same as they did in the raw game in itself
len(df[(df["is_a_package"]==True)&(df["package_and_game_view_is_identical"]==True)])

# checking if there are duplicates or not for unique values of recommendation_id and game_id
len(df[["recommendation_id","game_id"]].drop_duplicates())==len(df)

# Making sure the data is in its proper data type (for safe-measure)
if type(df["main_bar_list_of_similarities"][0]) != list:
    df["sidebar_list_of_similarities"] = [[str(item) for item in ast.literal_eval(lst)] for lst in tqdm(df["sidebar_list_of_similarities"])]
    df["main_bar_list_of_similarities"] = [[str(item) for item in ast.literal_eval(lst)] for lst in tqdm(df["main_bar_list_of_similarities"])]
    df["control_list_of_similarities"] = [[str(item) for item in ast.literal_eval(lst)] for lst in tqdm(df["control_list_of_similarities"])]
else:
    # Ensure each element in each list is a string
    df["sidebar_list_of_similarities"] = [[str(item) for item in lst] for lst in tqdm(df["sidebar_list_of_similarities"])]
    df["main_bar_list_of_similarities"] = [[str(item) for item in lst] for lst in tqdm(df["main_bar_list_of_similarities"])]
    df["control_list_of_similarities"] = [[str(item) for item in lst] for lst in tqdm(df["control_list_of_similarities"])]

df['recommendation_id'] = df['recommendation_id'].astype(str)
df['game_id'] = df['game_id'].astype(str)
df['timestamp_created'] = df['timestamp_created'].astype(int)
df["votes_up"]= df['votes_up'].astype(int)
df["voted_up"]= df['voted_up'].astype(bool)

def checking_length_of_comparison(df,columns_to_check = ["main_bar_list_of_similarities", 
                                                         "sidebar_list_of_similarities", 
                                                         "control_list_of_similarities"]):
    '''
    ensuring that each review_id has 10 similarities for each list sidebar, mainbar and control reviews
    Args:
    df (pandas dataframe): dataframe of our reviews with the relevant columns where each cell contains
    a list of similarities of comparison. 
    columns_to_check (list): a list of columns in which we want to ensure that each cell contains a list
    with a length of 10 similarities. 
    
    Returns:
    print statement with row and column containing the cells that do not contain a list of 10 similarities. 

    '''

    # Initialize an empty list to store rows and columns that don't meet the criterion
    mismatched_items = []

    # Iterate through each row and column to check the condition
    for col in tqdm(columns_to_check):
        for index, item in enumerate(df[col]):
            # Check if the item is a list and its length is 10
            if not isinstance(item, list) or len(item) != 10:
                mismatched_items.append((index, col))

    # Displaying the results
    if mismatched_items:
        for index, col in mismatched_items:
            print(f"Warning! Row {index} in column '{col}' does not have a list of length 10. Please double check!")
    else:
        print("All items in the specified columns are lists with a length of 10.")

        
checking_length_of_comparison(df,columns_to_check = ["main_bar_list_of_similarities", 
                                                     "sidebar_list_of_similarities", 
                                                     "control_list_of_similarities"])

def obtain_extended_dataframe_of_similarities(df, column_name, similarity_type):
    '''
    melts and extends data for each given similarity type (ie mainbar, sidebar, and control)
    
    args:
    df (pandas dataframe): dataframe with the columns of containing a list of similarities 
    for the corresponding similarity type
    column_name (str): the column name in which we wish to melt the data
    similarity_type (str): type of similarity we are comparing each review with
    
    returns:
    pandas dataframe of melted data that is ready for analysis. 
    '''
    def obtain_corresponding_feature(org_list):
        return([element for element in list(org_list) for _ in range(10)])
    similarities=[item for sublist in tqdm(list(df[column_name])) for item in (sublist)]
    recommendation_id=obtain_corresponding_feature((df["recommendation_id"]))
    order=[i+1 for i in range(10)]*len(df)
    game_id=obtain_corresponding_feature((df["game_id"]))
    score=obtain_corresponding_feature((df["score"]))
    votes_up=obtain_corresponding_feature((df["votes_up"]))
    voted_up=obtain_corresponding_feature((df["voted_up"]))
    timestamp_created=obtain_corresponding_feature((df["timestamp_created"]))
    appears_both_in_game_and_package=obtain_corresponding_feature((df["appears_both_in_game_and_package"]))
    is_a_package=obtain_corresponding_feature((df["is_a_package"]))
    package_and_game_view_is_identical=obtain_corresponding_feature((df["package_and_game_view_is_identical"]))
    package_review_but_displayed_different=obtain_corresponding_feature((df["package_review_but_displayed_different"]))
    similarity_type=[similarity_type for i in range(len(voted_up))]
    data=pd.DataFrame({
        "recommendation_id":recommendation_id,
        "order":order,
        "similarity_score":similarities,
        "similarity_type":similarity_type,
        "game_id":game_id,
        "score":score,
        "votes_up":votes_up,
        "voted_up":voted_up,
        "time_stampcreated":timestamp_created,
        "is_a_package":is_a_package,
        "appears_both_in_game_and_package":appears_both_in_game_and_package,
        "package_and_game_view_is_identical":package_and_game_view_is_identical,
        "package_review_but_displayed_different":package_review_but_displayed_different
        
    })
    return(data)

df_mainbar=obtain_extended_dataframe_of_similarities(df, "main_bar_list_of_similarities", "main bar")
df_sidebar=obtain_extended_dataframe_of_similarities(df, "sidebar_list_of_similarities", "sidebar")
df_control=obtain_extended_dataframe_of_similarities(df, "control_list_of_similarities", "control")
data=pd.concat([df_mainbar,df_sidebar,df_control]).reset_index(drop=True)

# checking for duplicates
duplicates = data.duplicated(keep=False)
(data[duplicates])

data["similarity_score"]=data["similarity_score"].astype(float)

data["votes_up"]=data["votes_up"].astype(int)

data.to_csv("../data/processed_data/data.csv.gz",index=False, compression='gzip')

number_duplicate_reviews = data.groupby('recommendation_id').size() == 60
print('{:,} reviews in total'.format(data.shape[0]//30))
print('{:,.0f} duplicates ({}%)'.format(number_duplicate_reviews.sum(), round(100*number_duplicate_reviews.mean(), 1)))
data = data.groupby('recommendation_id').filter(lambda x : x.shape[0] == 30).reset_index(drop=True)
print('{:,} reviews after excluding duplicates'.format(data.shape[0]//30))
game_length=set(list(data["game_id"]))
print("game length:", len(game_length))
