==> 06_summary_statistics.ipynb <==
import numpy as np
import pandas as pd
import types
import zipfile
import io
from tqdm.notebook import tqdm_notebook
import json
def imports():
    for name, val in globals().items():
        if isinstance(val, types.ModuleType):
            try:
                yield val.__name__, val.__version__
                pass
            except:
                yield val.__name__
                pass
            pass
        pass
    pass

list(imports())
pd.set_option('display.float_format', lambda x: '%.3f' % x)


# reading the CSV files
df=pd.read_csv("../data/raw_data/FPS_reviews.csv.zip",low_memory=False)
df_final=(df.drop_duplicates(["recommendation_id","game_id"])).reset_index(drop=True)

# getting the number of reviews with duplicates
game_length=len(set(list(df["game_id"])))
print(f"{len(df)} reviews across {game_length} games")

print("Total duplicates are",len(df)-len(df_final))

df=df_final

# review length after duplicates
game_length=len(set(list(df["game_id"])))
print(f"{len(df)} reviews across {game_length} games")

# getting some characteristic data
df["review"]=df["review"].astype(str)
df["review_words_length"]=df["review"].apply(lambda x: len(x.split(" ")))
df["review_words_vocab"]=df["review"].apply(lambda x: len(set(x.split(" "))))

df[["review_words_length","review_words_vocab","votes_up"]].describe()

df[["review_words_length","review_words_vocab","votes_up"]].median()

df["voted_up"].value_counts()

# Counting NA values in the 'voted_up' column
na_count = df['voted_up'].isna().sum()
print(f"Number of NA values in 'voted_up': {na_count}")

# getting other characteristic data not in the dataframe
raw_review_location = "../data/raw_data/" + 'reviews.txt.zip'
with zipfile.ZipFile(raw_review_location) as zf:
    with io.TextIOWrapper(zf.open('reviews.txt'), encoding='utf-8') as f:
        all_reviews_raw = f.readlines()
        pass
    pass
all_reviews = [json.loads(all_reviews_raw[xxx]) for xxx in tqdm_notebook(range(len(all_reviews_raw)))]


game_id, player_recommendtion_id_list, written_during_early_access_list = [], [], []
playtime_at_review_list, received_for_free_list,hidden_in_steam_china = [], [], []
for i in tqdm_notebook(range(len(all_reviews))):
    game_reviews = all_reviews[i]
    if len(game_reviews) > 0:
        for j in ((game_reviews)):
            review = j
            player_recommendtion_id_list.append(review["recommendationid"])
            hidden_in_steam_china.append(review["hidden_in_steam_china"])
            written_during_early_access_list.append(review["written_during_early_access"])
            received_for_free_list.append(review["received_for_free"])
            try:
                playtime_at_review_list.append(review["playtime_at_review"])
            except:
                playtime_at_review_list.append(0)
            game_id.append(review['game_id'])
        pass
    pass

df_other_characteristics = pd.DataFrame(
    list(zip(player_recommendtion_id_list, game_id, playtime_at_review_list,
             hidden_in_steam_china, written_during_early_access_list,
             received_for_free_list)),
    columns=['recommendation_id', 'game_id', 'playtime_at_review',
             'hidden_in_steam_china', 'written_during_early_access', 'received_for_free']
)
data_location="../data/supplemental_data/"
df_other_characteristics=(df_other_characteristics.drop_duplicates(["recommendation_id","game_id"])).reset_index(drop=True)
print(len(df_other_characteristics)==len(df))
df_other_characteristics.to_csv(f"{data_location}data_other_characteristics.csv.zip", index=False)

df_other_characteristics[["playtime_at_review"]].describe()/60

if df_other_characteristics["hidden_in_steam_china"].value_counts().sum()==len(df_other_characteristics):
    print(df_other_characteristics["hidden_in_steam_china"].value_counts())
    print(df_other_characteristics["hidden_in_steam_china"].value_counts()/len(df_other_characteristics))
else:
    print("values missing")
#     print(df_other_characteristics["hidden_in_steam_china"].value_counts()/len(df_other_characteristics))
#     print((len(df_other_characteristics)-df_other_characteristics["hidden_in_steam_china"].value_counts().sum)/len(df_other_characteristics))

if df_other_characteristics["received_for_free"].value_counts().sum()==len(df_other_characteristics):
    print(df_other_characteristics["received_for_free"].value_counts())
    print(df_other_characteristics["received_for_free"].value_counts()/len(df_other_characteristics))

if df_other_characteristics["written_during_early_access"].value_counts().sum()==len(df_other_characteristics):
    print(df_other_characteristics["written_during_early_access"].value_counts())
    print(df_other_characteristics["written_during_early_access"].value_counts()/len(df_other_characteristics))

row_counts = df.groupby('game_id').size()
average = row_counts.mean()
std_dev = row_counts.std()
median = row_counts.median()
print(f"Mean: {average}, Mdn:{median}, sd: {std_dev}")
